remove typo from rescue exception
